+++ spec/orm/models/person.rb
@@ -3,11 +3,15 @@ class Person < MassiveRecord::ORM::Table
     field :name
     field :email
     field :age, :integer
-    field :points, :integer, :default => 1, :column => :pts
     field :date_of_birth, :date
-    field :status, :boolean, :default => false
     field :addresses, :hash, :default => {}
   end
+
+  column_family :base do
+    field :points, :integer, :default => 1, :column => :pts
+    field :status, :boolean, :default => false
+  end
+
 
   references_one :boss, :class_name => "PersonWithTimestamp", :store_in => :info
   references_many :test_classes, :store_in => :info